feat:add HttpStatus

+++ src/main/java/com/practice/container/todo_api/controller/ToDoController.java
@@ -3,6 +3,8 @@ package com.practice.container.todo_api.controller;
 import com.practice.container.todo_api.entity.ToDo;
 import com.practice.container.todo_api.service.ToDoService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,28 +17,47 @@ public class ToDoController {
     private ToDoService toDoService;
 
     @GetMapping
-    public List<ToDo> getAllToDos() {
-        return toDoService.getAllToDos();
+    public ResponseEntity<List<ToDo>> getAllToDos() {
+        List<ToDo> todos = toDoService.getAllToDos();
+        return new ResponseEntity<>(todos, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public ToDo getToDoById(@PathVariable int id) {
-        return toDoService.getToDoById(id);
+    public ResponseEntity<ToDo> getToDoById(@PathVariable int id) {
+        ToDo todo = toDoService.getToDoById(id);
+        if (todo != null) {
+            return new ResponseEntity<>(todo, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @PostMapping
-    public void createToDo(@RequestBody ToDo todo) {
+    public ResponseEntity<ToDo> createToDo(@RequestBody ToDo todo) {
         toDoService.createToDo(todo);
+        return new ResponseEntity<>(todo, HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
-    public void updateToDoById(@PathVariable int id, @RequestBody ToDo todo) {
-        todo.setId(id);
-        toDoService.updateToDoById(todo);
+    public ResponseEntity<ToDo> updateToDoById(@PathVariable int id, @RequestBody ToDo todoDetails) {
+        ToDo existingToDo = toDoService.getToDoById(id);
+        if(existingToDo != null) {
+            todoDetails.setId(id);
+            toDoService.updateToDoById(todoDetails);
+            return new ResponseEntity<>(todoDetails, HttpStatus.OK);
+        }else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteToDoById(@PathVariable int id) {
-        toDoService.deleteToDoById(id);
+    public ResponseEntity<Void> deleteToDoById(@PathVariable int id) {
+        ToDo existingToDo = toDoService.getToDoById(id);
+        if (existingToDo != null){
+            toDoService.deleteToDoById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 }
